+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## 1.1.0 (2016-01-10)
 - Folders rearranged,
 - API of `DoctrineDatabaseBackup` class updated.
+
+## 1.2.0 (2016-02-21)
+- API of `DoctrineDatabaseBackup` class updated,
+- Added callback for simple database setup.